+++ R/map_layer_grid.R
@@ -93,7 +93,7 @@ add_grid <- function(
 	l[["lat"]] <- force(lat)
 	l[["polyline"]] <- force(polyline)
 
-	l <- resolve_data( data, l, "POINT" )
+	l <- resolve_data( data, l, c("POINT","MULTIPOINT") )
 
 	if ( !is.null(l[["data"]]) ) {
 		data <- l[["data"]]