+++ R/map_layer_scatterplot.R
@@ -114,7 +114,7 @@ add_scatterplot <- function(
 	l <- resolve_palette( l, palette )
 	l <- resolve_legend( l, legend )
 	l <- resolve_legend_options( l, legend_options )
-	l <- resolve_data( data, l, c( "POINT") )
+	l <- resolve_data( data, l, c( "POINT", "MULTIPOINT") )
 
 	if ( !is.null(l[["data"]]) ) {
 		data <- l[["data"]]