+++ R/map_layer_text.R
@@ -95,7 +95,7 @@ add_text <- function(
 	l <- resolve_palette( l, palette )
 	l <- resolve_legend( l, legend )
 	l <- resolve_legend_options( l, legend_options )
-	l <- resolve_data( data, l, "POINT")
+	l <- resolve_data( data, l, c("POINT","MULTIPOINT"))
 
 	if ( !is.null(l[["data"]]) ) {
 		data <- l[["data"]]