+++ src/RcppExports.cpp
@@ -48,16 +48,16 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_grid_geojson
-Rcpp::List rcpp_grid_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
-RcppExport SEXP _mapdeck_rcpp_grid_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
+Rcpp::List rcpp_grid_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_column);
+RcppExport SEXP _mapdeck_rcpp_grid_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
-    rcpp_result_gen = Rcpp::wrap(rcpp_grid_geojson(data, data_types, params, geometry_columns));
+    Rcpp::traits::input_parameter< std::string >::type geometry_column(geometry_columnSEXP);
+    rcpp_result_gen = Rcpp::wrap(rcpp_grid_geojson(data, data_types, params, geometry_column));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -90,7 +90,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_hexagon_geojson
-Rcpp::List rcpp_hexagon_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_hexagon_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_hexagon_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -98,7 +98,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_hexagon_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP
@@ -160,7 +160,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_path_geojson
-Rcpp::List rcpp_path_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_path_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_path_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -168,7 +168,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_path_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP
@@ -188,7 +188,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_pointcloud_geojson
-Rcpp::List rcpp_pointcloud_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_pointcloud_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_pointcloud_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -196,7 +196,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_pointcloud_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP
@@ -230,7 +230,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_polygon_geojson
-Rcpp::List rcpp_polygon_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_polygon_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_polygon_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -238,7 +238,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_polygon_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP
@@ -258,7 +258,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_scatterplot_geojson
-Rcpp::List rcpp_scatterplot_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_scatterplot_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_scatterplot_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -266,7 +266,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_scatterplot_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP
@@ -300,7 +300,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_screengrid_geojson
-Rcpp::List rcpp_screengrid_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_screengrid_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_screengrid_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -308,7 +308,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_screengrid_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP
@@ -342,7 +342,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_text_geojson
-Rcpp::List rcpp_text_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, Rcpp::StringVector geometry_columns);
+Rcpp::List rcpp_text_geojson(Rcpp::DataFrame data, Rcpp::List data_types, Rcpp::List params, std::string geometry_columns);
 RcppExport SEXP _mapdeck_rcpp_text_geojson(SEXP dataSEXP, SEXP data_typesSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -350,7 +350,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type data_types(data_typesSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type geometry_columns(geometry_columnsSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_text_geojson(data, data_types, params, geometry_columns));
     return rcpp_result_gen;
 END_RCPP