+++ src/pointcloud.cpp
@@ -16,7 +16,7 @@ Rcpp::List pointcloud_defaults(int n) {
 
 // [[Rcpp::export]]
 Rcpp::List rcpp_pointcloud_geojson( Rcpp::DataFrame data, Rcpp::List data_types,
-                            Rcpp::List params, Rcpp::StringVector geometry_columns ) {
+                            Rcpp::List params, std::string geometry_columns ) {
 
 	int data_rows = data.nrows();
 
@@ -25,7 +25,7 @@ Rcpp::List rcpp_pointcloud_geojson( Rcpp::DataFrame data, Rcpp::List data_types,
 	std::unordered_map< std::string, std::string > pointcloud_colours = mapdeck::pointcloud::pointcloud_colours;
 	Rcpp::StringVector pointcloud_legend = mapdeck::pointcloud::pointcloud_legend;
 
-	return spatialwidget::api::create_geojson(
+	return spatialwidget::api::create_geojson_downcast(
 		data,
 		data_types,
 		params,