+++ src/screengrid.cpp
@@ -14,7 +14,7 @@ Rcpp::List screengrid_defaults(int n) {
 
 // [[Rcpp::export]]
 Rcpp::List rcpp_screengrid_geojson( Rcpp::DataFrame data, Rcpp::List data_types,
-                            Rcpp::List params, Rcpp::StringVector geometry_columns ) {
+                            Rcpp::List params, std::string geometry_columns ) {
 
 	int data_rows = data.nrows();
 
@@ -22,7 +22,7 @@ Rcpp::List rcpp_screengrid_geojson( Rcpp::DataFrame data, Rcpp::List data_types,
 	std::unordered_map< std::string, std::string > screengrid_colours = mapdeck::screengrid::screengrid_colours;
 	Rcpp::StringVector screengrid_legend = mapdeck::screengrid::screengrid_legend;
 
-	return spatialwidget::api::create_geojson(
+	return spatialwidget::api::create_geojson_downcast(
 		data,
 		data_types,
 		params,