+++ NEWS.md
@@ -1,8 +1,8 @@
 # mapdeck 0.1.006
 
+* `MULTI`-geometry sf objects supported 
 * can use variables in place of string arguments
 * `highlight_colour` argument
 * all `add_*()` functions migrated to c++
-* `MULTIPOLYGON` not supported in `add_polygon()`
 * `layer_id` is optional to the user
 * `auto_highlight` argument